much of the header work: done
There is still a large chunk of work to do on the header, but i feel very good about what I currently have. Going to commit this now, and go back to it whenever I have finished it. Also: i may need to separate main.css and make a header.css in the future, but lets see how it goes. Check #front-end for an image of how this looks.